% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advr.r
\name{get_arduino}
\alias{get_arduino}
\title{Download an Arduino program file}
\usage{
get_arduino()
}
\value{
an .ino file for the Arduino microcontroller
}
\description{
Download an Arduino program file
}
\examples{
get_arduino()
}
