% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advr.r
\name{get_data}
\alias{get_data}
\title{Download a demo data file}
\usage{
get_data(exdir = "demo")
}
\arguments{
\item{exdir}{A character string with the name where the downloaded file is saved.}
}
\value{
A .zip data file downloaded in the destination directory.
}
\description{
Download a demo data file
}
\examples{
get_data()
}
