% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advr.r
\name{get_logger}
\alias{get_logger}
\title{Download the logger program}
\usage{
get_logger()
}
\value{
a .CR1 program file for the data logger
}
\description{
Download the logger program
}
\examples{
get_logger()
}
