% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idf.R
\name{idf}
\alias{idf}
\title{calculate Intensity-Duration-Frequency (IDF) quantiles and generate plot}
\usage{
idf(
  x_list,
  rp = c(2, 5, 10, 25, 50, 100),
  dur = c(1, 2, 5, 10, 24, 48),
  aggfun = "mean",
  dist = "gev",
  ...
)
}
\arguments{
\item{x_list}{a list of data tables containing time series data with columns "time", "value", and "cell_id"}

\item{rp}{vector of return periods for which IDF quantiles are calculated}

\item{dur}{vector of durations for which IDF quantiles are calculated (in hours)}

\item{aggfun}{aggregation function for calculating rolling statistics (default is "mean")}

\item{dist}{distribution to fit for IDF modeling (default is "gev")}
}
\value{
a list containing two elements:
  - 'data': a data table with columns for cell id, duration, return period, and corresponding IDF values
  - 'plot': a ggplot object visualizing the IDF curve
}
\description{
this function calculates IDF quantiles based on input data and parameters, and
generates a plot for visual representation of the IDF curve
}
\examples{
idf_result <- idf(x_list = my_data, rp = c(2, 5, 10, 25, 50, 100), dur = c(1, 2, 5, 10, 24, 48))
print(idf_result$data)
print(idf_result$plot)
}
