% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advr.r
\name{plot_adv}
\alias{plot_adv}
\title{plot raw data in day, hour, or other length}
\usage{
plot_adv(aa, sub_by = 1, suffix = "daily")
}
\arguments{
\item{aa}{pre processed raw data frame}

\item{sub_by}{sub figure by n days}

\item{suffix}{for the file name of the image}
}
\value{
figures in day, hour, or other length
}
\description{
plot raw data in day, hour, or other length
}
