% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{process_nc}
\alias{process_nc}
\title{import and process NetCDF data}
\usage{
process_nc(directory = "./data", pattern = ".nc", cell_ids)
}
\arguments{
\item{directory}{path to the directory containing the files}

\item{pattern}{file pattern to match NetCDF files}

\item{cell_ids}{vector of cell IDs to extract values}
}
\value{
a list of data tables split by cell_id
}
\description{
this function reads NetCDF files from a specified directory, extracts relevant
information for specified cell IDs, and returns returns a list of data tables with columns for time,
cell_id, and corresponding values
}
\examples{
process_nc(directory = "./data", pattern = ".nc", cell_ids = c(327, 328, 329))

}
