% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advr.r
\name{read_adv}
\alias{read_adv}
\title{Read and pre process the raw data}
\usage{
read_adv(path_raw, path_raw_header)
}
\arguments{
\item{path_raw}{A character string with the path of the raw data file.}

\item{path_raw_header}{A character string with path of header file for the raw data.}
}
\value{
A data frame with pre-processed raw data
}
\description{
Read and pre process the raw data
}
\examples{
get_data("demo")
adv_data <- read_adv(path_raw = "demo/data_sample.csv", 
    path_raw_header = "demo/data_header.csv")
}
