% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advr.r
\name{plot_adv}
\alias{plot_adv}
\title{plot raw data in day, hour, or other length}
\usage{
plot_adv(aa, sub_by = 1, suffix = "daily")
}
\arguments{
\item{aa}{A data frame with pre-processed raw data}

\item{sub_by}{A integer or NA indicating that the sub-figures are plotted by n days. if sub_by == NA, then plot the entire data set.}

\item{suffix}{A character string with the file name of the image}
}
\value{
Figures in day, hour, or other length.
}
\description{
plot raw data in day, hour, or other length
}
