% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advr.r
\name{read_adv}
\alias{read_adv}
\title{Read and pre process the raw data}
\usage{
read_adv(path_raw, path_raw_header)
}
\arguments{
\item{path_raw}{the path of the raw data file}

\item{path_raw_header}{the path of header file for the raw data}
}
\value{
a data frame with pre processed raw data
}
\description{
Read and pre process the raw data
}
\examples{
if (!dir.exists("demo")) dir.create("demo")
download.file("https://github.com/pzhaonet/advr/raw/master/inst2/data.zip", 
    "demo/data.zip")
unzip(zipfile = "demo/data.zip", exdir = "demo")
adv_data <- read_adv("demo/data_sample.csv", "demo/data_header.csv")
}
